backend/requirements.txt
fastapi
uvicorn
python-dotenv
firebase-admin
python-jose[cryptography]
psutil
pydantic
pydantic-settings
